Refactor and document usage as package
