+++ bin/wikidata-filter
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 var path = require('path')
-require(path.join(__dirname, '..', 'index.js'))
+require(path.join(__dirname, '..', 'index.js')).run()
