+++ lib/entity_parser.js
