+++ lib/entity_serializer.js
